update fix.sh to work with new bin/check output & change strategy
Doing all operations on the file system or zip files transparently lead
to lots of errors.  To reduce the complexity we opted for a new scheme:
For every operation involving a zip file, the zip file will be unpacked.
The operation can then happen in the file system, and at the end of all
operations the zip files can be recreated.