list rom dupes
This feature is somewhat useless because it's hard to decide which file
is the original, i.e. the one the user would like to keep.  The reported
duplicate is basically chosen randomly.  It does however give you a
correct count of how many duplicate roms there are and for what games.